Use a debugobj to get timings and log as they happen
Use a tracer object subclassing DBIx::Class::Storage::Statistics so we can log
each query as it starts (in case one hangs - we can see where it's got to), and
so we can calculate execution time of each query.

This opens the door to producing much more useful output - perhaps writing each
request's queries to a file, and adding routes to provide HTTP pages to list and
view them, and to find queries that took too long, etc.